requirements.txt
bokeh==2.0.2
Brotli==1.0.7
click==7.1.2
cloudpickle==1.4.1
contextvars==2.4
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
dask==2.18.0
distributed==2.18.0
findspark==1.4.1
Flask==1.1.2
Flask-Compress==1.5.0
fsspec==0.7.4
future==0.18.2
HeapDict==1.0.1
immutables==0.14
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
locket==0.2.0
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.18.5
packaging==20.4
pandas==1.0.4
partd==1.1.0
Pillow==7.1.2
plotly==4.8.1
psutil==5.7.0
py4j==0.10.7
pymongo==3.10.1
pyparsing==2.4.7
pyspark==2.4.6
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sortedcontainers==2.2.1
tblib==1.6.0
threadpoolctl==2.1.0
toolz==0.10.0
tornado==6.0.4
typing-extensions==3.7.4.2
Werkzeug==1.0.1
zict==2.0.0
